pyproject.toml
# this_file: pyproject.toml
# =============================================================================
# VEXY-LINES-RUN PACKAGE CONFIGURATION
# GUI desktop application for Vexy Lines style transfer.
# =============================================================================

# -----------------------------------------------------------------------------
# PROJECT METADATA
# -----------------------------------------------------------------------------
[project]
name = "vexy-lines-run"
description = "GUI desktop application for Vexy Lines style transfer"
readme = "README.md"
requires-python = ">=3.11,<3.14"
keywords = [
    "vexy-lines",
    "vector-art",
    "gui",
    "style-transfer",
    "creative-tools",
    "desktop-app",
]
dynamic = ["version"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]

dependencies = [
    "vexy-lines-py>=1.0.2",
    "vexy-lines-apy>=1.0.1",
    "loguru>=0.7.2",
    "customtkinter>=5.2.0",
    "CTkToolTip",
    "Pillow>=12.1.0",
    "tkinterdnd2>=0.4.0",
    "av>=17.0.0",
    "resvg-py>=0.2.0",
    "svglab>=0.1.0",
    "opencv-python>=4.8.0",
    "CTkMenuBarPlus>=0.1.0",
]

[[project.authors]]
name = "Fontlab Ltd."
email = "opensource@vexy.art"

[project.license]
text = "MIT"

[project.urls]
Documentation = "https://github.com/vexyart/vexy-lines/tree/main/vexy-lines-run#readme"
Issues = "https://github.com/vexyart/vexy-lines/issues"
Source = "https://github.com/vexyart/vexy-lines/tree/main/vexy-lines-run"

[project.scripts]
vexy-lines-run = "vexy_lines_run.__main__:main"

[project.gui-scripts]
vexy-lines-run = "vexy_lines_run.__main__:main"

# -----------------------------------------------------------------------------
# BUILD SYSTEM
# -----------------------------------------------------------------------------
[build-system]
requires = [
    "hatchling>=1.27.0",
    "hatch-vcs>=0.4.0",
]
build-backend = "hatchling.build"

# -----------------------------------------------------------------------------
# HATCH BUILD
# -----------------------------------------------------------------------------
[tool.hatch.build]
include = [
    "src/vexy_lines_run/py.typed",
]
exclude = ["**/__pycache__", "**/.pytest_cache", "**/.mypy_cache"]

[tool.hatch.build.targets.wheel]
packages = ["src/vexy_lines_run"]
reproducible = true

[tool.hatch.build.hooks.vcs]
version-file = "src/vexy_lines_run/_version.py"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.1.0"

[tool.hatch.metadata]
allow-direct-references = true

# -----------------------------------------------------------------------------
# HATCH ENVIRONMENTS
# -----------------------------------------------------------------------------
[tool.hatch.envs.default]
dependencies = [
    "ruff>=0.9.7",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "mkdocs>=1.6",
    "mkdocs-materialx>=2.0",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=src/vexy_lines_run --cov=tests {args:tests}"
type-check = "mypy src/vexy_lines_run tests"
lint = ["ruff check src/vexy_lines_run tests", "ruff format --check --respect-gitignore src/vexy_lines_run tests"]
fmt = ["ruff format --respect-gitignore src/vexy_lines_run tests", "ruff check --fix src/vexy_lines_run tests"]

[[tool.hatch.envs.all.matrix]]
python = ["3.11", "3.12", "3.13"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
    "ruff>=0.9.7",
    "mypy>=1.15.0",
]

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/vexy_lines_run tests}"
style = ["ruff check {args:.}", "ruff format --check --respect-gitignore {args:.}"]
fmt = ["ruff format --respect-gitignore {args:.}", "ruff check --fix {args:.}"]

[tool.hatch.envs.test]
dependencies = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[tool.hatch.envs.test.scripts]
test = "python -m pytest {args:tests}"
test-cov = "python -m pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=src/vexy_lines_run --cov=tests {args:tests}"

# -----------------------------------------------------------------------------
# COVERAGE
# -----------------------------------------------------------------------------
[tool.coverage.paths]
vexy_lines_run = ["src/vexy_lines_run", "*/vexy-lines-run/src/vexy_lines_run"]
tests = ["tests", "*/vexy-lines-run/tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
    "raise NotImplementedError",
]

[tool.coverage.run]
source_pkgs = ["vexy_lines_run", "tests"]
branch = true
parallel = true

# -----------------------------------------------------------------------------
# MYPY
# -----------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = [
    "customtkinter.*",
    "tkinterdnd2.*",
    "CTkMenuBarPlus.*",
    "CTkToolTip.*",
    "cv2.*",
    "av.*",
    "resvg.*",
    "svglab.*",
]
ignore_missing_imports = true

# -----------------------------------------------------------------------------
# PYTEST
# -----------------------------------------------------------------------------
[tool.pytest.ini_options]
addopts = "-v --durations=10 --tb=short"
console_output_style = "progress"
filterwarnings = ["ignore::DeprecationWarning"]
log_cli = true
log_cli_level = "INFO"
norecursedirs = [".*", "build", "dist", "venv", "__pycache__", "*.egg-info"]
python_classes = ["Test*"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
testpaths = ["tests"]

# -----------------------------------------------------------------------------
# RUFF
# -----------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "A", "ARG", "ASYNC", "B", "C", "DTZ", "E", "EM", "F", "FBT", "I", "ICN",
    "ISC", "LOG", "N", "PLC", "PLE", "PLR", "PLW", "PT", "PTH", "PYI", "RET",
    "RSE", "RUF", "S", "SIM", "T", "TCH", "TID", "UP", "W", "YTT",
]
ignore = [
    "B027", "C901", "FBT003", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
    "PLR1714", "PLW0603", "PT013", "PTH123", "PYI056", "S105", "S106",
    "S107", "S110", "SIM102",
]
unfixable = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["vexy_lines_run", "vexy_lines_api", "vexy_lines"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["PLR2004", "S101", "TID252"]
